+++ docs/main/changelog.rst
@@ -12,6 +12,7 @@ Development version
 .. - No features.
 - Add github badges to readme and docs index pages.
 - Update service names in services example miniapp.
+- Create engines & cars example miniapp.
 
 2.2.8
 -----